Struct rusoto_snowball::Address

pub struct Address {
    pub address_id: Option<String>,
    pub city: Option<String>,
    pub company: Option<String>,
    pub country: Option<String>,
    pub is_restricted: Option<bool>,
    pub landmark: Option<String>,
    pub name: Option<String>,
    pub phone_number: Option<String>,
    pub postal_code: Option<String>,
    pub prefecture_or_district: Option<String>,
    pub state_or_province: Option<String>,
    pub street_1: Option<String>,
    pub street_2: Option<String>,
    pub street_3: Option<String>,
}

The address that you want the Snowball or Snowballs associated with a specific job to be shipped to. Addresses are validated at the time of creation. The address you provide must be located within the serviceable area of your region. Although no individual elements of the Address are required, if the address is invalid or unsupported, then an exception is thrown.

Fields
address_id: Option<String>

The unique ID for an address.

city: Option<String>

The city in an address that a Snowball is to be delivered to.

company: Option<String>

The name of the company to receive a Snowball at an address.

country: Option<String>

The country in an address that a Snowball is to be delivered to.

is_restricted: Option<bool>

If the address you are creating is a primary address, then set this option to true. This field is not supported in most regions.

landmark: Option<String>

This field is no longer used and the value is ignored.

name: Option<String>

The name of a person to receive a Snowball at an address.

phone_number: Option<String>

The phone number associated with an address that a Snowball is to be delivered to.

postal_code: Option<String>

The postal code in an address that a Snowball is to be delivered to.

prefecture_or_district: Option<String>

This field is no longer used and the value is ignored.

state_or_province: Option<String>

The state or province in an address that a Snowball is to be delivered to.

street_1: Option<String>

The first line in a street address that a Snowball is to be delivered to.

street_2: Option<String>

The second line in a street address that a Snowball is to be delivered to.

street_3: Option<String>

The third line in a street address that a Snowball is to be delivered to.