Struct rusoto_snowball::CreateClusterRequest

pub struct CreateClusterRequest {
    pub address_id: String,
    pub description: Option<String>,
    pub forwarding_address_id: Option<String>,
    pub job_type: String,
    pub kms_key_arn: Option<String>,
    pub notification: Option<Notification>,
    pub resources: JobResource,
    pub role_arn: String,
    pub shipping_option: String,
    pub snowball_type: Option<String>,
}

Fields
address_id: String

The ID for the address that you want the cluster shipped to.>

description: Option<String>

An optional description of this specific cluster, for example Environmental Data Cluster-01.

forwarding_address_id: Option<String>

The forwarding address ID for a cluster. This field is not supported in most regions.

job_type: String

The type of job for this cluster. Currently, the only job type supported for clusters is LOCAL_USE.

kms_key_arn: Option<String>

The KmsKeyARN value that you want to associate with this cluster. KmsKeyARN values are created by using the CreateKey API action in AWS Key Management Service (AWS KMS).

notification: Option<Notification>

The Amazon Simple Notification Service (Amazon SNS) notification settings for this cluster.

resources: JobResource

The resources associated with the cluster job. These resources include Amazon S3 buckets and optional AWS Lambda functions written in the Python language.

role_arn: String

The RoleARN that you want to associate with this cluster. RoleArn values are created by using the CreateRole API action in AWS Identity and Access Management (IAM).

shipping_option: String

The shipping speed for each node in this cluster. This speed doesn't dictate how soon you'll get each Snowball Edge appliance, rather it represents how quickly each appliance moves to its destination while in transit. Regional shipping speeds are as follows:

In Australia, you have access to express shipping. Typically, appliances shipped express are delivered in about a day.

In the European Union (EU), you have access to express shipping. Typically, Snowball Edges shipped express are delivered in about a day. In addition, most countries in the EU have access to standard shipping, which typically takes less than a week, one way.

In India, Snowball Edges are delivered in one to seven days.

In the US, you have access to one-day shipping and two-day shipping.

snowball_type: Option<String>

The type of AWS Snowball appliance to use for this cluster. Currently, the only supported appliance type for cluster jobs is EDGE.