Struct rusoto_snowball::JobMetadata

pub struct JobMetadata {
    pub address_id: Option<String>,
    pub cluster_id: Option<String>,
    pub creation_date: Option<f64>,
    pub data_transfer_progress: Option<DataTransfer>,
    pub description: Option<String>,
    pub forwarding_address_id: Option<String>,
    pub job_id: Option<String>,
    pub job_log_info: Option<JobLogs>,
    pub job_state: Option<String>,
    pub job_type: Option<String>,
    pub kms_key_arn: Option<String>,
    pub notification: Option<Notification>,
    pub resources: Option<JobResource>,
    pub role_arn: Option<String>,
    pub shipping_details: Option<ShippingDetails>,
    pub snowball_capacity_preference: Option<String>,
    pub snowball_type: Option<String>,
}

Contains information about a specific job including shipping information, job status, and other important metadata. This information is returned as a part of the response syntax of the DescribeJob action.

Fields
address_id: Option<String>

The ID for the address that you want the Snowball shipped to.

cluster_id: Option<String>

The 39-character ID for the cluster, for example CID123e4567-e89b-12d3-a456-426655440000.

creation_date: Option<f64>

The creation date for this job.

data_transfer_progress: Option<DataTransfer>

A value that defines the real-time status of a Snowball's data transfer while the appliance is at AWS. This data is only available while a job has a JobState value of InProgress, for both import and export jobs.

description: Option<String>

The description of the job, provided at job creation.

forwarding_address_id: Option<String>

The ID of the address that you want a job shipped to, after it will be shipped to its primary address. This field is not supported in most regions.

job_id: Option<String>

The automatically generated ID for a job, for example JID123e4567-e89b-12d3-a456-426655440000.

job_log_info: Option<JobLogs>

Links to Amazon S3 presigned URLs for the job report and logs. For import jobs, the PDF job report becomes available at the end of the import process. For export jobs, your job report typically becomes available while the Snowball for your job part is being delivered to you.

job_state: Option<String>

The current status of the jobs.

job_type: Option<String>

The type of job.

kms_key_arn: Option<String>

The Amazon Resource Name (ARN) for the AWS Key Management Service (AWS KMS) key associated with this job. This ARN was created using the CreateKey API action in AWS KMS.

notification: Option<Notification>

The Amazon Simple Notification Service (Amazon SNS) notification settings associated with a specific job. The Notification object is returned as a part of the response syntax of the DescribeJob action in the JobMetadata data type.

resources: Option<JobResource>

An array of S3Resource objects. Each S3Resource object represents an Amazon S3 bucket that your transferred data will be exported from or imported into.

role_arn: Option<String>

The role ARN associated with this job. This ARN was created using the CreateRole API action in AWS Identity and Access Management (IAM).

shipping_details: Option<ShippingDetails>

A job's shipping information, including inbound and outbound tracking numbers and shipping speed options.

snowball_capacity_preference: Option<String>

The Snowball capacity preference for this job, specified at job creation. In US regions, you can choose between 50 TB and 80 TB Snowballs. All other regions use 80 TB capacity Snowballs.

snowball_type: Option<String>

The type of appliance used with this job.