Struct rusoto_snowball::ShippingDetails

pub struct ShippingDetails {
    pub inbound_shipment: Option<Shipment>,
    pub outbound_shipment: Option<Shipment>,
    pub shipping_option: Option<String>,
}

A job's shipping information, including inbound and outbound tracking numbers and shipping speed options.

Fields
inbound_shipment: Option<Shipment>

The Status and TrackingNumber values for a Snowball being delivered to the address that you specified for a particular job.

outbound_shipment: Option<Shipment>

The Status and TrackingNumber values for a Snowball being returned to AWS for a particular job.

shipping_option: Option<String>

The shipping speed for a particular job. This speed doesn't dictate how soon you'll get the Snowball from the job's creation date. This speed represents how quickly it moves to its destination while in transit. Regional shipping speeds are as follows:

In Australia, you have access to express shipping. Typically, Snowballs shipped express are delivered in about a day.

In the European Union (EU), you have access to express shipping. Typically, Snowballs shipped express are delivered in about a day. In addition, most countries in the EU have access to standard shipping, which typically takes less than a week, one way.

In India, Snowballs are delivered in one to seven days.

In the United States of America (US), you have access to one-day shipping and two-day shipping.