Struct rusoto_snowball::SnowballClient

pub struct SnowballClient { /* fields omitted */ }

A client for the Amazon Snowball API.

Methods

impl SnowballClient

pub fn new(region: Region) -> SnowballClient

Creates a client backed by the default tokio event loop.

The client will use the default credentials provider and tls client.

pub fn new_with<P, D>(
request_dispatcher: D,
credentials_provider: P,
region: Region
) -> SnowballClient where
P: ProvideAwsCredentials + Send + Sync + 'static,
P::Future: Send,
D: DispatchSignedRequest + Send + Sync + 'static,
D::Future: Send,