Crate rusoto_snowball

AWS Snowball is a petabyte-scale data transport solution that uses secure devices to transfer large amounts of data between your on-premises data centers and Amazon Simple Storage Service (Amazon S3). The Snowball commands described here provide access to the same functionality that is available in the AWS Snowball Management Console, which enables you to create and manage jobs for Snowball. To transfer data locally with a Snowball device, you'll need to use the Snowball client or the Amazon S3 API adapter for Snowball. For more information, see the User Guide.

If you're using the service, you're probably looking for SnowballClient and Snowball.

Structs

Address
The address that you want the Snowball or Snowballs associated with a specific job to be shipped to. Addresses are validated at the time of creation. The address you provide must be located within the serviceable area of your region. Although no individual elements of the Address are required, if the address is invalid or unsupported, then an exception is thrown.

CancelClusterRequest

CancelClusterResult

CancelJobRequest

CancelJobResult

ClusterListEntry
Contains a cluster's state, a cluster's ID, and other important information.

ClusterMetadata
Contains metadata about a specific cluster.

CompatibleImage
A JSON-formatted object that describes a compatible Amazon Machine Image (AMI), including the ID and name for a Snowball Edge AMI. This AMI is compatible with the device's physical hardware requirements, and it should be able to be run in an SBE1 instance on the device.

CreateAddressRequest

CreateAddressResult

CreateClusterRequest

CreateClusterResult

CreateJobRequest

CreateJobResult

DataTransfer
Defines the real-time status of a Snowball's data transfer while the device is at AWS. This data is only available while a job has a JobState value of InProgress, for both import and export jobs.

DescribeAddressRequest

DescribeAddressResult

DescribeAddressesRequest

DescribeAddressesResult

DescribeClusterRequest

DescribeClusterResult

DescribeJobRequest

DescribeJobResult

Ec2AmiResource
A JSON-formatted object that contains the IDs for an Amazon Machine Image (AMI), including the Amazon EC2 AMI ID and the Snowball Edge AMI ID. Each AMI has these two IDs to simplify identifying the AMI in both the AWS Cloud and on the device.

EventTriggerDefinition
The container for the EventTriggerDefinition$EventResourceARN.

GetJobManifestRequest

GetJobManifestResult

GetJobUnlockCodeRequest

GetJobUnlockCodeResult

GetSnowballUsageRequest

GetSnowballUsageResult

JobListEntry
Each JobListEntry object contains a job's state, a job's ID, and a value that indicates whether the job is a job part, in the case of an export job.

JobLogs
Contains job logs. Whenever Snowball is used to import data into or export data out of Amazon S3, you'll have the option of downloading a PDF job report. Job logs are returned as a part of the response syntax of the DescribeJob action in the JobMetadata data type. The job logs can be accessed for up to 60 minutes after this request has been made. To access any of the job logs after 60 minutes have passed, you'll have to make another call to the DescribeJob action.

For import jobs, the PDF job report becomes available at the end of the import process. For export jobs, your job report typically becomes available while the Snowball for your job part is being delivered to you.

The job report provides you insight into the state of your Amazon S3 data transfer. The report includes details about your job or job part for your records.

For deeper visibility into the status of your transferred objects, you can look at the two associated logs: a success log and a failure log. The logs are saved in comma-separated value (CSV) format, and the name of each log includes the ID of the job or job part that the log describes.

JobMetadata
Contains information about a specific job including shipping information, job status, and other important metadata. This information is returned as a part of the response syntax of the DescribeJob action.

JobResource
Contains an array of AWS resource objects. Each object represents an Amazon S3 bucket, an AWS Lambda function, or an Amazon Machine Image (AMI) based on Amazon EC2 that is associated with a particular job.

KeyRange
Contains a key range. For export jobs, a S3Resource object can have an optional KeyRange value. The length of the range is defined at job creation, and has either an inclusive BeginMarker, an inclusive EndMarker, or both. Ranges are UTF-8 binary sorted.

LambdaResource
Identifies

ListClusterJobsRequest

ListClusterJobsResult

ListClustersRequest

ListClustersResult

ListCompatibleImagesRequest

ListCompatibleImagesResult

ListJobsRequest

ListJobsResult

Notification
The Amazon Simple Notification Service (Amazon SNS) notification settings associated with a specific job. The Notification object is returned as a part of the response syntax of the DescribeJob action in the JobMetadata data type.

When the notification settings are defined during job creation, you can choose to notify based on a specific set of job states using the JobStatesToNotify array of strings, or you can specify that you want to have Amazon SNS notifications sent out for all job states with NotifyAll set to true.

S3Resource
Each S3Resource object represents an Amazon S3 bucket that your transferred data will be exported from or imported into. For export jobs, this object can have an optional KeyRange value. The length of the range is defined at job creation, and has either an inclusive BeginMarker, an inclusive EndMarker, or both. Ranges are UTF-8 binary sorted.

Shipment
The Status and TrackingNumber information for an inbound or outbound shipment.

ShippingDetails
A job's shipping information, including inbound and outbound tracking numbers and shipping speed options.

SnowballClient
A client for the Amazon Snowball API.

UpdateClusterRequest

UpdateClusterResult

UpdateJobRequest

UpdateJobResult

Enums

CancelClusterError
Errors returned by CancelCluster

CancelJobError
Errors returned by CancelJob

CreateAddressError
Errors returned by CreateAddress

CreateClusterError
Errors returned by CreateCluster

CreateJobError
Errors returned by CreateJob

DescribeAddressError
Errors returned by DescribeAddress

DescribeAddressesError
Errors returned by DescribeAddresses

DescribeClusterError
Errors returned by DescribeCluster

DescribeJobError
Errors returned by DescribeJob

GetJobManifestError
Errors returned by GetJobManifest

GetJobUnlockCodeError
Errors returned by GetJobUnlockCode

GetSnowballUsageError
Errors returned by GetSnowballUsage

ListClusterJobsError
Errors returned by ListClusterJobs

ListClustersError
Errors returned by ListClusters

ListCompatibleImagesError
Errors returned by ListCompatibleImages

ListJobsError
Errors returned by ListJobs

UpdateClusterError
Errors returned by UpdateCluster

UpdateJobError
Errors returned by UpdateJob

Traits

Snowball
Trait representing the capabilities of the Amazon Snowball API. Amazon Snowball clients implement this trait.